Задание 5. Добавила Contact Helper.
