+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/appmanager/ContactHelper.java
@@ -1,0 +1,66 @@
+package ru.stqa.pft.addressbook.appmanager;
+
+import org.openqa.selenium.Alert;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import ru.stqa.pft.addressbook.model.ContactData;
+
+
+import static org.testng.Assert.assertTrue;
+
+public class ContactHelper {
+  private WebDriver driver;
+  private boolean acceptNextAlert = true;
+  public ContactHelper(WebDriver driver) {
+    this.driver = driver;
+  }
+  public void submitContactCreation() {
+    driver.findElement(By.xpath("//div[@id='content']/form/input[21]")).click();
+  }
+
+  public void fillContactForm(ContactData contactData) {
+    driver.findElement(By.name("firstname")).click();
+    driver.findElement(By.name("firstname")).clear();
+    driver.findElement(By.name("firstname")).sendKeys(contactData.firstname());
+    driver.findElement(By.name("lastname")).click();
+    driver.findElement(By.name("lastname")).clear();
+    driver.findElement(By.name("lastname")).sendKeys(contactData.lastname());
+    driver.findElement(By.name("mobile")).click();
+    driver.findElement(By.name("mobile")).clear();
+    driver.findElement(By.name("mobile")).sendKeys(contactData.mobile());
+    driver.findElement(By.name("email")).click();
+    driver.findElement(By.name("email")).clear();
+    driver.findElement(By.name("email")).sendKeys(contactData.email());
+  }
+
+  public void initContactCreation() {
+    driver.findElement(By.linkText("add new")).click();
+  }
+
+  public void acceptContactDeletion() {
+    assertTrue(closeAlertAndGetItsText().matches("^Delete 1 addresses[\\s\\S]$"));
+  }
+
+  public void deleteSelectedContacts() {
+    driver.findElement(By.xpath("//input[@value='Delete']")).click();
+  }
+
+  public void selectContact() {
+    driver.findElement(By.name("selected[]")).click();
+  }
+  private String closeAlertAndGetItsText() {
+    try {
+      Alert alert = driver.switchTo().alert();
+      String alertText = alert.getText();
+      if (acceptNextAlert) {
+        alert.accept();
+      } else {
+        alert.dismiss();
+      }
+      return alertText;
+    } finally {
+      acceptNextAlert = true;
+    }
+  }
+
+}
