+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/appmanager/NavigationHelper.java
@@ -2,6 +2,11 @@ package ru.stqa.pft.addressbook.appmanager;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+import java.time.Duration;
+
 
 public class NavigationHelper {
   private WebDriver driver;
@@ -12,5 +17,10 @@ public class NavigationHelper {
 
   public void gotoGroupPage() {
     driver.findElement(By.linkText("groups")).click();
+  }
+  public void returnToHomepage() {
+    WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+    wait.until(ExpectedConditions.visibilityOfElementLocated(By.linkText("home")));
+    driver.findElement(By.linkText("home")).click();
   }
 }
