+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/tests/ContactCreationTests.java
@@ -7,10 +7,10 @@ public class ContactCreationTests extends TestBase {
 
   @Test
   public void testContactCreation() throws Exception {
-    app.initContactCreation();
-    app.fillContactForm(new ContactData("Ivan", "Ivanov", "89274223344", "ivanivanov@gmail.com"));
-    app.submitContactCreation();
-    app.returnToHomepage();
+    app.getContactHelper().initContactCreation();
+    app.getContactHelper().fillContactForm(new ContactData("Ivan", "Ivanov", "89274223344", "ivanivanov@gmail.com"));
+    app.getContactHelper().submitContactCreation();
+    app.getNavigationHelper().returnToHomepage();
   }
 
 }
