+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/tests/ContactDeletionTests.java
@@ -1,15 +1,16 @@
 package ru.stqa.pft.addressbook.tests;
 
+import org.openqa.selenium.support.ui.Wait;
 import org.testng.annotations.Test;
 
 public class ContactDeletionTests extends TestBase {
 
   @Test
   public void testContactDeletion() throws Exception {
-    app.selectContact();
-    app.deleteSelectedContacts();
-    app.acceptDeletion();
-    app.returnToHomepage();
+    app.getContactHelper().selectContact();
+    app.getContactHelper().deleteSelectedContacts();
+    app.getContactHelper().acceptContactDeletion();
+    app.getNavigationHelper().returnToHomepage();
   }
 
 }
