+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/tests/GroupDeletionTests.java
@@ -1,6 +1,7 @@
 package ru.stqa.pft.addressbook.tests;
 
 import org.testng.annotations.Test;
+import ru.stqa.pft.addressbook.model.ContactData;
 
 public class GroupDeletionTests extends TestBase {
